+++ returns/hkt.py
@@ -31,3 +31,33 @@ class Kind(Generic[_InstanceType, _TypeArgTypes]):
         return super().__class_getitem__(  # type: ignore
             type_params[:real_args],
         )
+
+
+def dekind(
+    kind: Kind[_InstanceType, _TypeArgTypes],
+) -> _InstanceType:
+    """
+    Turns ``Kind[IO, int]`` type into real ``IO[int]`` type.
+
+    Should be used when you are left with accidential ``Kind`` instance
+    when you really want to have the real type.
+
+    Works with type arguments of any length.
+
+    We use a custom ``mypy`` plugin to make sure types are correct.
+    Otherwise, it is currently impossible to properly type this.
+
+    In runtime it just returns the passed argument, nothing really happens:
+
+    .. code:: python
+
+      >>> from returns.io import IO
+      >>> from returns.hkt import Kind
+
+      >>> container: Kind[IO, int] = IO(1)
+      >>> assert dekind(container) is container
+
+    However, please, do not use this function
+    unless you know exactly why do you need it.
+    """
+    return kind  # type: ignore